requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
